MDEXP-638 - Exporting shared instance without holdings
